requirements.txt
# Django
Django==3.1.5
# Django Extensions
django-extensions==3.1.0
# Flake8
flake8==3.8.4
# Widget Tweaks
 django-widget-tweaks==1.4.8
 mercadopago==1.2.0

 # Django REST Framework
djangorestframework==3.12.2
django-cors-headers==3.6.0
django-filter==2.4.0
drf-yasg==1.20.0
djangorestframework-simplejwt==4.6.0

django-environ==0.4.5  # https://github.com/joke2k/django-environ

psycopg2-binary==2.8.6

gunicorn==20.1.0
whitenoise==5.2.0
django-cors-headers==3.6.0
factory-boy==3.2.0